4.0.1-ET2: Исправлена сериализация `instructions` в `Response`
https://github.com/AdrianMiska/openai-kotlin/tree/feature/Responses-API

+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/responses/ResponseTool.kt
@@ -118,7 +118,7 @@ public sealed interface ResponseTool {
          * Defaults to 0
          */
         @SerialName("score_threshold")
-        val scoreThreshold: Int? = null,
+        val scoreThreshold: Double? = null,
     )
